EPBDS-9041 fix casting from Float to BigInteger. Note: The number 3.4E+38 can be represented as float and impossible to represented it as long. But it can be perfectly represented as BigInteger

+++ DEV/org.openl.core/src/org/openl/meta/DoubleValue.java
@@ -716,7 +716,7 @@ public class DoubleValue extends ExplanationNumberValue<DoubleValue> implements 
     }
 
     public static BigInteger cast(DoubleValue x, BigInteger y) {
-        return BigInteger.valueOf(x.longValue());
+        return BigDecimal.valueOf(x.doubleValue()).toBigInteger();
     }
 
     public static BigDecimal cast(DoubleValue x, BigDecimal y) {
